break: Change logger to build to lib.
Because of the way everything is setup if we wanted to import from a
submodule we ended up having to import from /dist which just didn't sit
right with me. Because of this I decided to change things up so we build
to lib making the imports a little nicer to read e.g. `import {log} from 'logger/lib/utils'`

Some alternatives to this were

1. cp the package.json into the dist directory and yarn link the
   package... This had some other issues and it also meant we'd have to
   do this for every shared lib which isn't scalable